restructure Controller files to be readable
